requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.20
langchain-openai==0.1.8
langchain-anthropic==0.1.23
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.1
